Added TFCond_Zoomed fix for third-person props
When players are in third-person mode and attempt to scope in with a
Sniper Rifle, their scoped-in view height is offset; players need to
compensate by aiming a good amount over their target.

This fix forces the player back into first-person mode while zoomed in.

Also, moved the conditions into their own methods for organization.